feat: refactor raffle screen layout by consolidating narrow and wide layouts into a single implementation for improved code maintainability
